feat: refactor order confirmation page and add OrderConfirmationClient component
